Implemented the process of get metadata from builders.
1.For CharOctetLength and CharacterMaximumLength, the type should be
long.
2.For infinite loop, it was caused in getMetaData stage.
TableBuilder.getMetaData() invokes ColumnBuilder.getMetaData(), then
ColumnBuilder invokes TableBuilder.getMetaData() ...….. We leave the
table of column to be assigned after table metadata is constructed.
3.

+++ Madz.DatabaseMetaData/src/main/java/net/madz/db/core/meta/immutable/ColumnMetaData.java
@@ -34,7 +34,7 @@ public interface ColumnMetaData<SMD extends SchemaMetaData<SMD, TMD, CMD, FMD, I
 
     String getDefaultValue();
 
-    Integer getCharOctetLength();
+    Long getCharOctetLength();
 
     /** Position of this column in physical table layout (first column=1, ...) */
     Short getOrdinalPosition();
